Gave getChasisSpeeds an input
I measured the robot wheel diameter in inches and took the rpm of the motor using the getVelocity function. I then converted the diameter from inches to meters and then multiplied by pi to get the circumference and then multiplied that by the RPM to get the meters per second of the left and right side of the robot which I then inputted into the new DifferentialDriveWheelSpeeds function to get a variable that I inputted into getChasisSpeeds.