Fixed subsystems, comments, and updated a change in constants
Got rid of the drive subsystem that was inside of CANRoller subsystem for some reason (all the code inside was a copy of what was in auto generator). Fixed some comments and one change  in constants that was supposed to be done when I gave chasisSpeeds an input wasn't commited for some reason so im commiting it now.